==> 14. sensor_and_source_PAC_viz.ipynb <==
## PLOTTING PAC SOURCES

Algorithm to plot sources:
1. Iterate through subjects, for each:
    - get PAC data
    - transpose
    - impute NaNs
    - aggregate
2. Stack PAC estimates along subjects axis
3. Average across ph, amp and sub


4. Load any morphed source file:
    - average time domain to get 1 point
    - replace data with PAC data
    - rename subject to "fsaverage_bem"
    - PLOT as regular source estimate
    - (optional) add ROI, if needed

ADDITIONAL:
- sensor-level PAC
- sensor- and source-level TFR

import copy
import numpy as np
import os
from utils import check_paths

import mne
from mne.stats import permutation_cluster_1samp_test

import scipy
from scipy.stats import zscore
from scipy.sparse import coo_matrix, save_npz

import matplotlib.pyplot as plt
import matplotlib.patches as patches
%matplotlib qt

eeg_data_dir = 'D:\\BonoKat\\research project\\# study 1\\eeg_data\\set'
groups = ['Y']
tasks = ['_BL', '_MAIN'] # ['_BL', '_MAIN']
task_stages = ['_plan', '_go']
# block_names = ['_baseline', '_adaptation']

# Load adjacency matrix
src_fname = 'D:\\BonoKat\\research project\\# study 1\\mri_data\\fs_output\\freesurfer\\sub_dir\\Y\\fsaverage_bem\\bem\\fsaverage-ico4-src.fif'
src = mne.read_source_spaces(src_fname)
# src.plot(subjects_dir='D:\\BonoKat\\research project\\# study 1\\mri_data\\fs_output\\freesurfer\\sub_dir\\Y')
source_adjacency = mne.spatial_src_adjacency(src)
adj = source_adjacency.tocsr()  # Ensure adjacency is CSR format for fast indexing
print('adjacency shape:', source_adjacency.shape)

# Load one stc file as a reference
stc_path = os.path.join('''D:\\BonoKat\\research project\\# study 1\\eeg_data\\set\\Y\\s1_pac_sub01\\preproc\\analysis\\source\\morphed_stcs\\_MAIN\\_plan\\_baseline\\s1_pac_sub01_MAIN_plan_baseline-stc-lcmv_epoch_000_morphed-lh.stc''')
stc_ref = mne.read_source_estimate(stc_path, subject='s1_pac_sub01') # any STC will serve as a template
stc_ref_mean = stc_ref.copy().mean()
stc_pac = stc_ref_mean.copy()
stc_pac.subject = "fsaverage_bem"
stc_pac

# Plot PAC MI values

views = {'lateral': ["lateral", "medial"],
        'dor-vent': ["dorsal", "ventral"]}

for group in groups:
    source_group_dir = os.path.join(eeg_data_dir, f'{group} group', 'sources')
    subs = os.listdir(os.path.join(eeg_data_dir, group))
    figs_save_path = os.path.join(eeg_data_dir, f'{group} group', 'sources', 'figs')
    check_paths(figs_save_path)

    

    for task in tasks:
        if task == '_BL':
            block_names = ['']
        else:
            block_names = ['_baseline', '_adaptation']

        for task_stage in task_stages:
            for block_name in block_names:
                print(f'Processing {group} group, {task}{task_stage}{block_name}...')

                pac_list = []
                for sub_name in subs:
                    analysis_dir = os.path.join(eeg_data_dir, group, sub_name, 'preproc', 'analysis')
                    pac_save_path = os.path.join(analysis_dir, 'source', 'PAC')
                    pac_data = np.load(os.path.join(pac_save_path, f"PAC_MI_SOURCE_{sub_name[-5:]}{task}{task_stage}{block_name}.npy"))
                    pac_data_t = np.transpose(pac_data, (1, 0, 2))

                    ### NAN Imputation for PAC Matrices ###
                    pac_imputed = pac_data_t.copy()

                    # Step 1: Detect vertices with any NaN in their 20×20 PAC
                    nan_mask = np.isnan(pac_data_t).any(axis=(1, 2))  # shape (5124,)

                    if nan_mask.any() == True:
                        # print(f"Found {nan_mask.sum()} vertices with NaN values.")

                        # Step 2: Impute NaNs from neighbors
                        for vtx in np.where(nan_mask)[0]:
                            neighbors = adj[[vtx]].indices
                            valid_neighbors = [n for n in neighbors if not nan_mask[n]]
                            # Average PAC matrices from neighbors
                            pac_imputed[vtx] = np.nanmean(pac_data_t[valid_neighbors], axis=0)

                    # pac_zscore = zscore(pac_imputed, axis=0, nan_policy='omit')
                    pac_list.append(pac_imputed)

                # Stack them along a new first axis (subject axis)
                pac_all = np.stack(pac_list, axis=0)
                # print('PAC array shape:', pac_all.shape)
                pac_map = pac_all.mean(axis=(0, 2, 3))
                # print('Averaged PAC map shape:', pac_map.shape)
                stc_pac._data = pac_map[:, None]   # (n_vertices_total, 1)
                # print('STC data shape:', stc_pac.data.shape)
                # Save morphed stc
                stc_pac.save(os.path.join(source_group_dir, f"PAC_for_STATS_{group}{task}{task_stage}{block_name}_fsaverage"), overwrite=True)

                clim = dict(
                    kind="value",
                    lims=[
                        np.nanmin(pac_map), 
                        (np.nanmax(pac_map) - np.nanmin(pac_map)) / 2 + np.nanmin(pac_map), 
                        np.nanmax(pac_map)
                        ],   # [min, mid, max]
                    )

                for key, value in views.items():
                    title = f"ALL_SUBS_{group}{task}{task_stage}{block_name}_PAC_MI_{key}"

                    brain = stc_pac.plot(subject="fsaverage_bem",
                                initial_time=0.0,
                                clim=clim,
                                cortex="low_contrast",
                                colormap='plasma', # 'plasma' for raw MI, 'cool' for zscored MI
                                hemi='both',
                                views=value, # "dorsal", "ventral"
                                background='white',
                                title = title
                                )

                    brain.save_image(os.path.join(figs_save_path, title + '.png'))


PLOTTING STATS

# Plot significant PAC

# Significant conditions:
# FTT (BL) planning
# De-CRAT (MAIN) planning baseline

group_save_path = os.path.join(eeg_data_dir, f'{group} group')
pac_stats_save_path = os.path.join(group_save_path, 'source_pac_stats')
sig_conditions = ['Y_BL_plan', 'Y_MAIN_plan_baseline']
p_thresh = 0.05

for sig_cond in sig_conditions:

    # Load stats data
    T_obs = np.load(os.path.join(pac_stats_save_path, f"PAC_MI_SOURCE_{sig_cond}_freqs_ave_T_obs.npy"))
    clusters = np.load(os.path.join(pac_stats_save_path, f"PAC_MI_SOURCE_{sig_cond}_freqs_ave_clusters.npy"), allow_pickle=True)
    cluster_p_values = np.load(os.path.join(pac_stats_save_path, f"PAC_MI_SOURCE_{sig_cond}_freqs_ave_cluster_p_values.npy"))

    # Combine significant cluster masks
    sig_mask = np.zeros_like(T_obs, dtype=bool)
    for cluster, p_val in zip(clusters, cluster_p_values):
        if p_val <= p_thresh:
            sig_mask |= cluster.astype(bool)
    sig_mask = sig_mask[:, None]
    title = f"Significant PAC - {sig_cond}"

    sig_cond = [sig_cond + '_' if sig_cond == 'Y_BL_plan' else sig_cond][0]
    group, task, task_stage, block_name = sig_cond.split('_')
    block_name = ['_' + block_name if block_name != '' else block_name][0]

    for sub_name in subs:

        pac_list = []

        analysis_dir = os.path.join(eeg_data_dir, group, sub_name, 'preproc', 'analysis')
        pac_save_path = os.path.join(analysis_dir, 'source', 'PAC')
        pac_data = np.load(os.path.join(pac_save_path, f"PAC_MI_SOURCE_{sub_name[-5:]}{'_' + task}{'_' + task_stage}{block_name}.npy"))
        pac_data_t = np.transpose(pac_data, (1, 0, 2))

        ### NAN Imputation for PAC Matrices ###
        pac_imputed = pac_data_t.copy()

        # Step 1: Detect vertices with any NaN in their 20×20 PAC
        nan_mask = np.isnan(pac_data_t).any(axis=(1, 2))  # shape (5124,)

        if nan_mask.any() == True:
            # print(f"Found {nan_mask.sum()} vertices with NaN values.")

            # Step 2: Impute NaNs from neighbors
            for vtx in np.where(nan_mask)[0]:
                neighbors = adj[[vtx]].indices
                valid_neighbors = [n for n in neighbors if not nan_mask[n]]
                # Average PAC matrices from neighbors
                pac_imputed[vtx] = np.nanmean(pac_data_t[valid_neighbors], axis=0)

        # pac_zscore = zscore(pac_imputed, axis=0, nan_policy='omit')
        pac_list.append(pac_imputed)

    # Stack them along a new first axis (subject axis)
    pac_all = np.stack(pac_list, axis=0)
    # print('PAC array shape:', pac_all.shape)
    pac_map = pac_all.mean(axis=(0, 2, 3))
    # print('Averaged PAC map shape:', pac_map.shape)
    stc_pac._data = pac_map[:, None]   # (n_vertices_total, 1)
    # print('STC data shape:', stc_pac.data.shape)
    
    # Copy the STC to avoid modifying the original
    stc_sig = copy.deepcopy(stc_pac)
    stc_sig.data[~sig_mask] = 0.0

    clim = dict(
        kind="value",
        lims=[
            np.nanmin(pac_map), 
            (np.nanmax(pac_map) - np.nanmin(pac_map)) / 2 + np.nanmin(pac_map), 
            np.nanmax(pac_map)
            ],   # [min, mid, max]
        )
    # view = [["dorsal", "lateral"] if sig_cond == 'Y_BL_plan_' else ["dorsal", "parietal"]]
    brain = stc_sig.plot(subject="fsaverage_bem",
                initial_time=0.0,
                clim=clim,
                cortex="low_contrast",
                colormap='hot', # 'plasma' for raw MI, 'cool' for zscored MI
                hemi='both',
                views=[["dorsal", "lateral"] if sig_cond == 'Y_BL_plan_' else ["dorsal", "parietal"]][0],
                background='white',
                title = title,
                )
    brain.save_image(os.path.join(figs_save_path, title + '.png'))

____________________________

PLOTTING PAC MI ON SENSOR LEVEL

eeg_data_dir = 'D:\\BonoKat\\research project\\# study 1\\eeg_data\\set'
groups = ['Y']
tasks = ['_BL', '_MAIN'] # ['_BL', '_MAIN']
task_stages = ['_plan', '_go']
# block_names = ['_baseline', '_adaptation']

epochs_path = os.path.join(eeg_data_dir, group, 's1_pac_sub01', 'preproc', 'analysis', 's1_pac_sub01_BL_epochs_go-epo.fif')
epochs = mne.read_epochs(epochs_path, preload=True)
evoked = epochs.average()
evoked.data.shape


for group in groups:
    subs = os.listdir(os.path.join(eeg_data_dir, group))
    sensor_figs_path = os.path.join(eeg_data_dir, f'{group} group', 'sensors')
    check_paths(sensor_figs_path)
    figs_save_path = os.path.join(sensor_figs_path, 'figs', 'PAC')
    check_paths(figs_save_path)
    

    for task in tasks:
        if task == '_BL':
            block_names = ['']
        else:
            block_names = ['_baseline', '_adaptation']

        for task_stage in task_stages:
            for block_name in block_names:
                print(f'Processing {group} group, {task}{task_stage}{block_name}...')

                pac_list = []
                for sub_name in subs:

                    analysis_dir = os.path.join(eeg_data_dir, group, sub_name, 'pac_results')
                    pac_data = np.load(os.path.join(analysis_dir, f"pac_mi_TOPO_{sub_name[-5:]}{task}{task_stage}{block_name}.npy"))
                    pac_data_t = np.transpose(pac_data, (1, 0, 2))
                    pac_list.append(pac_data_t)

                # Stack them along a new first axis (subject axis)
                pac_all = np.stack(pac_list, axis=0)
                pac_all_ave = np.mean(pac_all, axis=(0, 2, 3))
                print(pac_all_ave.shape)
                evoked.data = pac_all_ave


                fig, ax = plt.subplots(figsize=(6, 6))  # increase figure size here
                im, cn = mne.viz.plot_topomap(
                    evoked.data,
                    evoked.info,
                    ch_type="eeg",
                    sensors=True,
                    names=evoked.ch_names,
                    cmap="plasma",
                    vlim=(evoked.data.min(), evoked.data.max()),
                    axes=ax,
                    show=False
                )

                # add colorbar
                plt.colorbar(im, ax=ax)
                plt.title(f"PAC MI: {group}{task}{task_stage}{block_name}")
                plt.show()
                plt.savefig(os.path.join(figs_save_path, f"PAC_MI_topo{task}{task_stage}{block_name}.png"))
                plt.close()

                    

__________________________

Sensor-level TF and PSD by bands

eeg_data_dir = 'D:\\BonoKat\\research project\\# study 1\\eeg_data\\set'
groups = ['Y']
tasks = ['_BL', '_MAIN'] # ['_BL', '_MAIN']
task_stages = ['_plan', '_go']
sub_name = 'ALL_subs'

for group in groups:

    group_save_path = f'D:\\BonoKat\\research project\\# study 1\\eeg_data\\set\\{group} group'

    subs_dir = os.path.join(eeg_data_dir, group)
    subs = os.listdir(subs_dir)
    sensor_figs_path = os.path.join(eeg_data_dir, f'{group} group', 'sensors')
    tfr_figs_save_path = os.path.join(sensor_figs_path, 'figs', 'TF')
    psd_figs_save_path = os.path.join(sensor_figs_path, 'figs', 'PSD_bands')
    check_paths(tfr_figs_save_path, psd_figs_save_path)

    for task in tasks:
        if task == '_BL':
            block_names = ['']
        else:
            block_names = ['_baseline', '_adaptation']

        for task_stage in task_stages:
            for block_name in block_names:
                print(f'Processing {group} group, {task}{task_stage}{block_name}...')


                epochs_all_subs = mne.read_epochs(os.path.join(
                    group_save_path, f"{group}_{sub_name}{task}_epochs{task_stage}{block_name}_ALL-epo.fif"
                    ), preload=True)
                #! Change baseline for prestim period (if not applied yet)
                epochs_all_subs.apply_baseline(baseline=(-0.5, -0.001))
 
                freqs = np.logspace(*np.log10([4.0, epochs_all_subs.info['lowpass']]), num=40)
                n_cycles = freqs / 2.0  # different number of cycle per frequency
                baseline_tf = (epochs_all_subs.times.min(), -0.01)

                ### COMPUTE POWER TFR AND ITC ###
                power, itc = epochs_all_subs.compute_tfr(
                    method="morlet",
                    freqs=freqs,
                    n_cycles=n_cycles,
                    average=True,
                    return_itc=True,
                    decim=3,
                    n_jobs=-1
                )
                power.apply_baseline(baseline=baseline_tf, mode="zscore")

                # plot ITC and Power
                itc_fig = itc.plot(combine='mean',
                                    title=f'ITC: {group}{task}{task_stage}{block_name}',
                                    cmap='autumn')[0]
                power_all_fig = power.plot(mode="zscore",
                                            title=f'Averaged Power: {task}{task_stage}{block_name}',
                                            combine='mean', cmap='summer')[0]

                ax1, ax2 = itc_fig.axes[0], power_all_fig.axes[0]   # main TFR axis

                quadmesh_itc = ax1.collections[0]  # the heatmap
                quadmesh_itc.set_clim(vmin=-0.15, vmax=0.15)

                quadmesh_power = ax2.collections[0]  # the heatmap
                quadmesh_power.set_clim(vmin=-4, vmax=4)

                ### COMPUTE POWER SPECTRAL DENSITY BY BAND ###
                spectrum = epochs_all_subs.compute_psd()
                bands = {'Theta (4-8 Hz)': (4, 8),
                        'Alpha (8-12 Hz)': (8, 12),
                        'Beta (12-30 Hz)': (12, 30),
                        'Gamma (30-50 Hz)': (30, 50),
                        'High gamma (50-80 Hz)': (50, 80)}

                fig_psd = spectrum.plot_topomap(bands=bands,
                                                vlim="joint",
                                                normalize=True,
                                                cmap='summer')
                fig_psd.suptitle(f'PSD Topomap: {group}{task}{task_stage}{block_name}')


                # save ITC and Power plots
                itc_fig.savefig(os.path.join(tfr_figs_save_path,
                                             f"{group}_{sub_name}{task}{task_stage}{block_name}_itc.png"),
                                             dpi=300)
                power_all_fig.savefig(os.path.join(tfr_figs_save_path,
                                                   f"{group}_{sub_name}{task}{task_stage}{block_name}_power_all.png"),
                                                   dpi=300)
                fig_psd.savefig(os.path.join(psd_figs_save_path, f"{group}_{sub_name}{task}{task_stage}{block_name}_psd_topomap.png"), dpi=300)

# plt.close('all')


________________________

Source-level TF modulations by bands

bands = {'theta': (4, 8),
        'alpha': (8, 12),
        'beta': (12, 30),
        'gamma': (30, 80)}

eeg_data_dir = 'D:\\BonoKat\\research project\\# study 1\\eeg_data\\set'
sub_name = 's1_pac_sub64'
group = 'Y'
task = '_MAIN' # ['_BL', '_MAIN']
task_stage = '_go'
block_name = '_adaptation'
band = 'theta'

analysis_dir = os.path.join(eeg_data_dir, group, sub_name, 'preproc', 'analysis')
stcs_path = os.path.join(analysis_dir, 'source', 'morphed_stcs', task, task_stage, block_name) 

for stc_file in os.listdir(stcs_path):
    if stc_file.endswith('-rh.stc'): # MNE will load both hemispheres anyway
        stc_path = os.path.join(stcs_path, stc_file)
        stc = mne.read_source_estimate(stc_path, subject=sub_name)
        stc.crop(tmin=0, tmax=None)


        stcs_bands_path = os.path.join(stcs_path, band)
        check_paths(stcs_bands_path)
        l_freq, h_freq = bands[band]
        stc._data = stc.data.astype(float)
        stc.filter(l_freq, h_freq, n_jobs=-1, method='fir', fir_window='hamming')
        stc.save(os.path.join(stcs_bands_path, stc_file[:-4] + f'_{band}-stc'), overwrite=True)

# Filter source estimates by frequency bands and save
# DONE

for group in groups:
    subs = os.listdir(os.path.join(eeg_data_dir, group))    

    for task in tasks:
        if task == '_BL':
            block_names = ['']
        else:
            block_names = ['_baseline', '_adaptation']

        for task_stage in task_stages:
            for block_name in block_names:
                print(f'Filtering sources: {group}{task}{task_stage}{block_name}...')

                for sub_name in subs:                    
                    analysis_dir = os.path.join(eeg_data_dir, group, sub_name, 'preproc', 'analysis')
                    stcs_path = os.path.join(analysis_dir, 'source', 'morphed_stcs', task, task_stage, block_name) 

                    for stc_file in os.listdir(stcs_path):
                        if stc_file.endswith('-rh.stc'): # MNE will load both hemispheres anyway
                            stc_path = os.path.join(stcs_path, stc_file)
                            stc = mne.read_source_estimate(stc_path, subject=sub_name)
                            stc.crop(tmin=0, tmax=None)

                            for band in bands.keys():
                                stcs_bands_path = os.path.join(stcs_path, band)
                                check_paths(stcs_bands_path)
                                l_freq, h_freq = bands[band]
                                stc._data = stc.data.astype(float)
                                stc.filter(l_freq, h_freq, n_jobs=-1, method='fir', fir_window='hamming')
                                stc.save(os.path.join(stcs_bands_path, stc_file[:-4] + f'_{band}-stc'), overwrite=True)


# Optimize band-specific storage: Average and visualize stcs per band, then remove single epochs
# - Save sub's STCs in .npy format
# - Remove original .stc files
# - save averaged STC
# - save movies


for group in groups:
    subs = os.listdir(os.path.join(eeg_data_dir, group))
    figs_save_path = os.path.join(f'D:\\BonoKat\\research project\\# study 1\\eeg_data\\set\\figures\\{group}', 'stc_movies')
    check_paths(figs_save_path)

    for sub_name in subs:
        analysis_dir = os.path.join(eeg_data_dir, group, sub_name, 'preproc', 'analysis')

        for task in tasks:
            if task == '_BL':
                block_names = ['']
            else:
                block_names = ['_baseline', '_adaptation']

            for task_stage in task_stages:
                for block_name in block_names:
                    for band in bands.keys():
                        stcs_path = os.path.join(analysis_dir, 'source', 'morphed_stcs', task, task_stage, block_name, band)
                        
                        print(f'Filtering sources: {group}{task}{task_stage}{block_name}_{band}...')

                        stcs = []
                        stcs_data = []

                        for stc_file in os.listdir(stcs_path):
                            if stc_file.endswith('-rh.stc'): # MNE will load both hemispheres anyway
                                stc_path = os.path.join(stcs_path, stc_file)
                                stc = mne.read_source_estimate(stc_path, subject=sub_name)
                                stcs.append(stc)
                                stcs_data.append(stc.data)
                                #!!!!!!!!!!! removing the files !!!!!!!!!!
                                os.remove(stc_path) # remove -rh file
                                os.remove(os.path.join(stc_path[:-7] + '-lh.stc')) # remove -lh file

                        stc_ave = stc.copy()
                        stcs_data = np.array(stcs_data)
                        stc_ave.data = np.mean(stcs_data, axis=0)
                        stc_ave.subject = "fsaverage_bem"

                        clim = dict(
                            kind="value",
                            lims=[
                                np.nanmin(stc_ave.data), 
                                (np.nanmax(stc_ave.data) - np.nanmin(stc_ave.data)) / 2 + np.nanmin(stc_ave.data), 
                                np.nanmax(stc_ave.data)
                                ],   # [min, mid, max]
                            )
                        fig_title = f'{sub_name}_average STC for {task[1:]}{task_stage}{block_name}, {band} band'
                        file_name = f'{sub_name}_STCs{task}{task_stage}{block_name}_{band}'

                        brain = stc_ave.plot(subject="fsaverage_bem",
                                    initial_time=0.0,
                                    clim=clim,
                                    cortex="low_contrast",
                                    colormap='seismic',
                                    hemi='both',
                                    views=["dorsal", "lateral", "medial"],
                                    background='white',
                                    title = fig_title,
                                    )

                        brain.save_movie(os.path.join(figs_save_path, file_name + '_AVE.mp4'), time_dilation=10.0)
                        stc_ave.save(os.path.join(stcs_path, file_name + '_AVE'))
                        np.save(os.path.join(stcs_path, file_name + '-epo-data.npy'), stc_ave.data)
                        brain.close()



eeg_data_dir = 'D:\\BonoKat\\research project\\# study 1\\eeg_data\\set'
groups = ['Y']
tasks = ['_BL', '_MAIN']
task_stages = ['_plan', '_go']
bands = {
    'theta': (4, 8),
    'alpha': (8, 12),
    'beta': (12, 30),
    'gamma': (30, 100)
}

# Average STCs per band across all participants

for group in groups:
    subs = os.listdir(os.path.join(eeg_data_dir, group))
    figs_save_path = os.path.join(f'D:\\BonoKat\\research project\\# study 1\\eeg_data\\set\\{group} group\\sources\\figs', 'stc_movies')
    stc_save_path = os.path.join(f'D:\\BonoKat\\research project\\# study 1\\eeg_data\\set\\{group} group\\sources', 'stcs_by_bands')
    check_paths(figs_save_path)
    check_paths(stc_save_path)

    for task in tasks:
        if task == '_BL':
            block_names = ['']
        else:
            block_names = ['_baseline', '_adaptation']

        for task_stage in task_stages:
            for block_name in block_names:

                for band in bands.keys():
                    print(f'Analysing sources: {group}{task}{task_stage}{block_name}_{band}...')
                    ave_stcs = []
                    ave_stcs_data = []

                    for sub_name in subs:

                        analysis_dir = os.path.join(eeg_data_dir, group, sub_name, 'preproc', 'analysis')
                        stcs_path = os.path.join(analysis_dir, 'source', 'morphed_stcs', task, task_stage, block_name, band)
                        stc_file = f'{sub_name}_STCs{task}{task_stage}{block_name}_{band}_AVE-rh.stc'
                        stc = mne.read_source_estimate(os.path.join(stcs_path, stc_file), subject=sub_name)
                        ave_stcs.append(stc)
                        ave_stcs_data.append(stc.data)

                    ave_stcs_all = stc.copy()
                    ave_stcs_data = np.array(ave_stcs_data)
                    ave_stcs_all.data = np.mean(ave_stcs_data, axis=0)
                    ave_stcs_all.subject = "fsaverage_bem"

                    clim = dict(
                        kind="value",
                        lims=[
                            np.nanmin(ave_stcs_all.data), 
                            (np.nanmax(ave_stcs_all.data) - np.nanmin(ave_stcs_all.data)) / 2 + np.nanmin(ave_stcs_all.data), 
                            np.nanmax(ave_stcs_all.data)
                            ],   # [min, mid, max]
                        )
                    fig_title = f'{group}_average STC for {task[1:]}{task_stage}{block_name}, {band} band'
                    file_name = f'{group}_STCs{task}{task_stage}{block_name}_{band}'

                    brain = ave_stcs_all.plot(subject="fsaverage_bem",
                                initial_time=0.0,
                                clim=clim,
                                cortex="low_contrast",
                                colormap='seismic',
                                hemi='both',
                                views=["dorsal", "lateral", "medial"],
                                background='white',
                                title = fig_title,
                                )

                    brain.save_movie(os.path.join(figs_save_path, file_name + '_GRAND_AVE.mp4'), time_dilation=10.0)
                    ave_stcs_all.save(os.path.join(stc_save_path, file_name + '_GRAND_AVE'), overwrite=True)
                    brain.close()



STOP HERE

______
Plotting drafts

Check for zeros and nans in averaged scts

stcs_path = os.path.join('D:\\BonoKat\\research project\\# study 1\\eeg_data\\set\\Y group\\sources\\stcs_by_bands\\Y_STCs_MAIN_plan_baseline_beta_GRAND_AVE-rh.stc')
stcs = mne.read_source_estimate(stcs_path, subject='fsaverage_bem')
stcs

clim = dict(
    kind="value",
    lims=[
        np.nanmin(stcs.data), 
        (np.nanmax(stcs.data) - np.nanmin(stcs.data)) / 2 + np.nanmin(stcs.data), 
        np.nanmax(stcs.data)
        ],   # [min, mid, max]
    )

stcs.plot(subject="fsaverage_bem", clim=clim, hemi='both', views=["dorsal", "lateral", "medial"], background='white', colormap='seismic')

print(np.nanmin(stcs.data[:, 100]), np.nanmax(stcs.data[:, 100]))

np.where(stcs.data == 0)

stcs.data[649][46]

___________

eeg_data_dir = 'D:\\BonoKat\\research project\\# study 1\\eeg_data\\set'
group = 'Y'
figs_save_path = os.path.join(eeg_data_dir, f'{group} group', 'sources', 'figs')
check_paths(figs_save_path)


sub_name = 's1_pac_sub07'
analysis_dir = os.path.join(eeg_data_dir, group, sub_name, 'preproc', 'analysis')
task = '_MAIN'
task_stage = '_plan'
block_name = '_baseline'

stcs_path = os.path.join(analysis_dir, 'source', 'morphed_stcs', task, task_stage, block_name)

stcs = []
stcs_data = []

tmin = 0.0
tmax = 0.495

for stc_file in os.listdir(stcs_path):
    if stc_file.endswith('-rh.stc'): # MNE will load both hemispheres anyway
        stc_path = os.path.join(stcs_path, stc_file)
        stc = mne.read_source_estimate(stc_path, subject=sub_name)
        stc.crop(tmin=tmin, tmax=tmax)
        stcs.append(stc)
        stcs_data.append(stc.data)

source_array = np.stack(stcs_data, axis=0)
source_array.shape # (epochs x vertices x time)

# mri_data_dir = f'D:\\BonoKat\\research project\\# study 1\\mri_data\\fs_output\\freesurfer\\sub_dir\\Y'
# fname_fsaverage_src = os.path.join(mri_data_dir, "fsaverage_bem", "bem", "fsaverage-ico4-src.fif")

# eeg_data_dir = os.path.join('D:\\BonoKat\\research project\\# study 1\\eeg_data\\set', group)
# src_to = mne.read_source_spaces(fname_fsaverage_src)

# # Morph the source estimates to the fsaverage space
# morph = mne.compute_source_morph(
#     stc,
#     subject_from=sub_name,
#     subject_to="fsaverage_bem",
#     src_to=src_to,
#     subjects_dir=mri_data_dir,
#     smooth=5 # small smoothing to avoid artifacts in the morphed data
# )

# stc_fsaverage = morph.apply(stc)
# stc.plot()
# stc_fsaverage.plot()
# stc_fsaverage

analysis_dir = os.path.join(eeg_data_dir, group, sub_name, 'preproc', 'analysis')
pac_save_path = os.path.join(analysis_dir, 'source', 'PAC')
pac_data = np.load(os.path.join(pac_save_path, f"PAC_MI_SOURCE_{sub_name[-5:]}{task}{task_stage}{block_name}.npy"))

pac_data_t = np.transpose(pac_data, (1, 0, 2))

### NAN Imputation for PAC Matrices ###
pac_imputed = pac_data_t.copy()

# Step 1: Detect vertices with any NaN in their 20×20 PAC
nan_mask = np.isnan(pac_data_t).any(axis=(1, 2))  # shape (5124,)

if nan_mask.any() == True:
    print(f"Found {nan_mask.sum()} vertices with NaN values.")

    # Step 2: Impute NaNs from neighbors
    for vtx in np.where(nan_mask)[0]:
        neighbors = adj[[vtx]].indices
        valid_neighbors = [n for n in neighbors if not nan_mask[n]]
        # Average PAC matrices from neighbors
        pac_imputed[vtx] = np.nanmean(pac_data_t[valid_neighbors], axis=0)

pac_zscore = zscore(pac_imputed, axis=0, nan_policy='omit')
pac_map = pac_zscore.mean(axis=(1, 2))
pac_map_clean = np.nan_to_num(pac_map, nan=1, posinf=0.0, neginf=0.0)
stc_ref = stcs[0]  # any one of your loaded STCs
stc_ref_mean = stc_ref.copy().mean()
stc_pac = stc_ref_mean.copy()
stc_pac._data = pac_map_clean[:, None]   # (n_vertices_total, 1)
stc_pac.subject = "fsaverage_bem"
stc_pac

clim = dict(
    kind="value",
    lims=[np.nanmin(pac_map), (np.nanmax(pac_map) - np.nanmin(pac_map)) / 2 + np.nanmin(pac_map), np.nanmax(pac_map)],   # [min, mid, max]
    
)

views = {'lateral': ["lateral", "medial"],
         'dor-vent': ["dorsal", "ventral"]}

for key, value in views.items():
    title = f"{sub_name[-5:]}{task}{task_stage}{block_name}_z-score_PAC_{key}"

    brain = stc_pac.plot(subject="fsaverage_bem",
                initial_time=0.0,
                clim=clim,
                cortex="low_contrast",
                colormap='plasma', # 'plasma' for raw MI, 'cool' for zscored MI
                hemi='both',
                views=value, # "dorsal", "ventral"
                background='white',
                title = title
                )

    # brain.save_image(os.path.join(figs_save_path, title))

## Plotting with ROI labels, if needed

# Select a few ROIs (names depend on your parcellation)
labels = mne.read_labels_from_annot('fsaverage_bem', parc='aparc')
roi_names = ["postcentral-lh", "precentral-lh", "superiorfrontal-lh"]
rois = [lab for lab in labels if lab.name in roi_names]

brain = stc_pac.plot(
    subject="fsaverage_bem",
    hemi="both",
    views=["lateral", "medial"],
    cortex="bone",
    clim=clim,
    colormap="plasma",
    initial_time=0.0,
)

# Add them in different colors
colors = ["yellow", "cyan", "magenta"]
for roi, color in zip(rois, colors):
    brain.add_label(roi, borders=True, color=color)

_____
